requirements.txt
google-cloud-language
requests
flask_sqlalchemy
psycopg2